# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Incident and Employee Management System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type EmployeeStatus =
  (typeof EmployeeStatus)[keyof typeof EmployeeStatus];

export const EmployeeStatus = {
  active: "active",
  inactive: "inactive",
} as const;

export interface Employee {
  id: string;
  name: string;
  email: string;
  department: string;
  role: string;
  phone?: string;
  status: EmployeeStatus;
  createdAt: string;
  updatedAt: string;
}

export type CreateEmployeeBodyStatus =
  (typeof CreateEmployeeBodyStatus)[keyof typeof CreateEmployeeBodyStatus];

export const CreateEmployeeBodyStatus = {
  active: "active",
  inactive: "inactive",
} as const;

export interface CreateEmployeeBody {
  name: string;
  email: string;
  department: string;
  role: string;
  phone?: string;
  status?: CreateEmployeeBodyStatus;
}

export type UpdateEmployeeBodyStatus =
  (typeof UpdateEmployeeBodyStatus)[keyof typeof UpdateEmployeeBodyStatus];

export const UpdateEmployeeBodyStatus = {
  active: "active",
  inactive: "inactive",
} as const;

export interface UpdateEmployeeBody {
  name?: string;
  email?: string;
  department?: string;
  role?: string;
  phone?: string;
  status?: UpdateEmployeeBodyStatus;
}

export type IncidentPriority =
  (typeof IncidentPriority)[keyof typeof IncidentPriority];

export const IncidentPriority = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export type IncidentStatus =
  (typeof IncidentStatus)[keyof typeof IncidentStatus];

export const IncidentStatus = {
  open: "open",
  investigating: "investigating",
  resolved: "resolved",
} as const;

export interface Incident {
  id: string;
  ticketNumber: string;
  title: string;
  description: string;
  priority: IncidentPriority;
  status: IncidentStatus;
  category?: string;
  assignedTo?: string | null;
  assignedToName?: string | null;
  reportedBy: string;
  reportedByName: string;
  resolvedAt?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type CreateIncidentBodyPriority =
  (typeof CreateIncidentBodyPriority)[keyof typeof CreateIncidentBodyPriority];

export const CreateIncidentBodyPriority = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export interface CreateIncidentBody {
  title: string;
  description: string;
  priority: CreateIncidentBodyPriority;
  category?: string;
  assignedTo?: string;
  reportedBy: string;
  reportedByName: string;
}

export type UpdateIncidentBodyPriority =
  (typeof UpdateIncidentBodyPriority)[keyof typeof UpdateIncidentBodyPriority];

export const UpdateIncidentBodyPriority = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export type UpdateIncidentBodyStatus =
  (typeof UpdateIncidentBodyStatus)[keyof typeof UpdateIncidentBodyStatus];

export const UpdateIncidentBodyStatus = {
  open: "open",
  investigating: "investigating",
  resolved: "resolved",
} as const;

export interface UpdateIncidentBody {
  title?: string;
  description?: string;
  priority?: UpdateIncidentBodyPriority;
  status?: UpdateIncidentBodyStatus;
  category?: string;
  assignedTo?: string;
}

export type UpdateIncidentStatusBodyStatus =
  (typeof UpdateIncidentStatusBodyStatus)[keyof typeof UpdateIncidentStatusBodyStatus];

export const UpdateIncidentStatusBodyStatus = {
  open: "open",
  investigating: "investigating",
  resolved: "resolved",
} as const;

export interface UpdateIncidentStatusBody {
  status: UpdateIncidentStatusBodyStatus;
  note?: string;
  updatedBy?: string;
}

export interface AddIncidentLogBody {
  action: string;
  note?: string;
  updatedBy: string;
}

export interface IncidentLog {
  id: string;
  incidentId: string;
  action: string;
  note?: string;
  updatedBy: string;
  timestamp: string;
}

export interface IncidentListResponse {
  data: Incident[];
  total: number;
  page: number;
  limit: number;
}

export interface DashboardStats {
  totalIncidents: number;
  openIncidents: number;
  investigatingIncidents: number;
  resolvedIncidents: number;
  criticalIncidents: number;
  highPriorityIncidents: number;
  totalEmployees: number;
  activeEmployees: number;
}

export interface PriorityCount {
  priority: string;
  count: number;
}

export interface StatusCount {
  status: string;
  count: number;
}

export type ListEmployeesParams = {
  search?: string;
  department?: string;
  role?: string;
};

export type ListIncidentsParams = {
  search?: string;
  status?: ListIncidentsStatus;
  priority?: ListIncidentsPriority;
  assignedTo?: string;
  page?: number;
  limit?: number;
};

export type ListIncidentsStatus =
  (typeof ListIncidentsStatus)[keyof typeof ListIncidentsStatus];

export const ListIncidentsStatus = {
  open: "open",
  investigating: "investigating",
  resolved: "resolved",
} as const;

export type ListIncidentsPriority =
  (typeof ListIncidentsPriority)[keyof typeof ListIncidentsPriority];

export const ListIncidentsPriority = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export type GetRecentIncidentsParams = {
  limit?: number;
};