# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Incident and Employee Management System API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: employees
    description: Employee management
  - name: incidents
    description: Incident tracking
  - name: dashboard
    description: Dashboard statistics

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /employees:
    get:
      operationId: listEmployees
      tags: [employees]
      summary: List all employees
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: department
          in: query
          schema:
            type: string
        - name: role
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of employees
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Employee"
    post:
      operationId: createEmployee
      tags: [employees]
      summary: Create a new employee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEmployeeBody"
      responses:
        "201":
          description: Employee created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"

  /employees/{id}:
    get:
      operationId: getEmployee
      tags: [employees]
      summary: Get employee by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Employee record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateEmployee
      tags: [employees]
      summary: Update an employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEmployeeBody"
      responses:
        "200":
          description: Employee updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
    delete:
      operationId: deleteEmployee
      tags: [employees]
      summary: Delete an employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  /incidents:
    get:
      operationId: listIncidents
      tags: [incidents]
      summary: List all incidents
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [open, investigating, resolved]
        - name: priority
          in: query
          schema:
            type: string
            enum: [low, medium, high, critical]
        - name: assignedTo
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of incidents
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IncidentListResponse"
    post:
      operationId: createIncident
      tags: [incidents]
      summary: Create a new incident
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateIncidentBody"
      responses:
        "201":
          description: Incident created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Incident"

  /incidents/{id}:
    get:
      operationId: getIncident
      tags: [incidents]
      summary: Get incident by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Incident record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Incident"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateIncident
      tags: [incidents]
      summary: Update an incident
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateIncidentBody"
      responses:
        "200":
          description: Incident updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Incident"
    delete:
      operationId: deleteIncident
      tags: [incidents]
      summary: Delete an incident
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  /incidents/{id}/status:
    patch:
      operationId: updateIncidentStatus
      tags: [incidents]
      summary: Update incident status with a log entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateIncidentStatusBody"
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Incident"

  /incidents/{id}/logs:
    get:
      operationId: getIncidentLogs
      tags: [incidents]
      summary: Get history logs for an incident
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Incident history logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/IncidentLog"
    post:
      operationId: addIncidentLog
      tags: [incidents]
      summary: Add a log entry to an incident
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddIncidentLogBody"
      responses:
        "201":
          description: Log entry added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IncidentLog"

  /dashboard/stats:
    get:
      operationId: getDashboardStats
      tags: [dashboard]
      summary: Get dashboard summary statistics
      responses:
        "200":
          description: Dashboard statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"

  /dashboard/recent-incidents:
    get:
      operationId: getRecentIncidents
      tags: [dashboard]
      summary: Get recent incidents for dashboard
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Recent incidents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Incident"

  /dashboard/priority-breakdown:
    get:
      operationId: getPriorityBreakdown
      tags: [dashboard]
      summary: Count of incidents grouped by priority
      responses:
        "200":
          description: Priority breakdown
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PriorityCount"

  /dashboard/status-breakdown:
    get:
      operationId: getStatusBreakdown
      tags: [dashboard]
      summary: Count of incidents grouped by status
      responses:
        "200":
          description: Status breakdown
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StatusCount"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]

    Employee:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        department:
          type: string
        role:
          type: string
        phone:
          type: string
        status:
          type: string
          enum: [active, inactive]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, name, email, department, role, status, createdAt, updatedAt]

    CreateEmployeeBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        department:
          type: string
        role:
          type: string
        phone:
          type: string
        status:
          type: string
          enum: [active, inactive]
          default: active
      required: [name, email, department, role]

    UpdateEmployeeBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        department:
          type: string
        role:
          type: string
        phone:
          type: string
        status:
          type: string
          enum: [active, inactive]

    Incident:
      type: object
      properties:
        id:
          type: string
        ticketNumber:
          type: string
        title:
          type: string
        description:
          type: string
        priority:
          type: string
          enum: [low, medium, high, critical]
        status:
          type: string
          enum: [open, investigating, resolved]
        category:
          type: string
        assignedTo:
          type: string
          nullable: true
        assignedToName:
          type: string
          nullable: true
        reportedBy:
          type: string
        reportedByName:
          type: string
        resolvedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, ticketNumber, title, description, priority, status, reportedBy, reportedByName, createdAt, updatedAt]

    CreateIncidentBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        priority:
          type: string
          enum: [low, medium, high, critical]
        category:
          type: string
        assignedTo:
          type: string
        reportedBy:
          type: string
        reportedByName:
          type: string
      required: [title, description, priority, reportedBy, reportedByName]

    UpdateIncidentBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        priority:
          type: string
          enum: [low, medium, high, critical]
        status:
          type: string
          enum: [open, investigating, resolved]
        category:
          type: string
        assignedTo:
          type: string

    UpdateIncidentStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [open, investigating, resolved]
        note:
          type: string
        updatedBy:
          type: string
      required: [status]

    AddIncidentLogBody:
      type: object
      properties:
        action:
          type: string
        note:
          type: string
        updatedBy:
          type: string
      required: [action, updatedBy]

    IncidentLog:
      type: object
      properties:
        id:
          type: string
        incidentId:
          type: string
        action:
          type: string
        note:
          type: string
        updatedBy:
          type: string
        timestamp:
          type: string
          format: date-time
      required: [id, incidentId, action, updatedBy, timestamp]

    IncidentListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Incident"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [data, total, page, limit]

    DashboardStats:
      type: object
      properties:
        totalIncidents:
          type: integer
        openIncidents:
          type: integer
        investigatingIncidents:
          type: integer
        resolvedIncidents:
          type: integer
        criticalIncidents:
          type: integer
        highPriorityIncidents:
          type: integer
        totalEmployees:
          type: integer
        activeEmployees:
          type: integer
      required:
        - totalIncidents
        - openIncidents
        - investigatingIncidents
        - resolvedIncidents
        - criticalIncidents
        - highPriorityIncidents
        - totalEmployees
        - activeEmployees

    PriorityCount:
      type: object
      properties:
        priority:
          type: string
        count:
          type: integer
      required: [priority, count]

    StatusCount:
      type: object
      properties:
        status:
          type: string
        count:
          type: integer
      required: [status, count]